# src/generated/types/Pokemon.ts
//
// THIS FILE HAS BEEN AUTOGENERATED. DO NOT MODIFY.
//

import * as t from "io-ts"

export const PokemonSchema = JSON.parse(`{
  "type": "object",
  "properties": {
    "id": {
      "type": "number"
    },
    "name": {
      "type": "string"
    }
  }
}`);

export const Pokemon = t.partial({
  id: t.number,
  name: t.string
});

export type Pokemon = t.TypeOf<typeof Pokemon>