# src/generated/types/Pokemons.ts
//
// THIS FILE HAS BEEN AUTOGENERATED. DO NOT MODIFY.
//

import * as t from "io-ts"
import { Pokemon } from "./Pokemon"

export const PokemonsSchema = JSON.parse(`{
  "type": "array",
  "items": {
    "type": "object",
    "properties": {
      "id": {
        "type": "number"
      },
      "name": {
        "type": "string"
      }
    }
  }
}`);

export const Pokemons = t.array(Pokemon);

export type Pokemons = t.TypeOf<typeof Pokemons>